+++ src/company_gui.cpp
@@ -24,6 +24,7 @@
 #include "strings_func.h"
 #include "timer/timer_game_economy.h"
 #include "dropdown_type.h"
+#include "dropdown_common_type.h"
 #include "tilehighlight_func.h"
 #include "company_base.h"
 #include "core/geometry_func.hpp"
